Previous change to pwiz RAMPAdapter to provide activationMethod values didn't properly null terminate the activationMethod string.  Adding a cast in call to min() to satisfy gcc.
Former-commit-id: cacafb9fe849dfc7055e22939a8dd9bf7e6be60f

+++ pwiz/data/msdata/RAMPAdapter.cpp
@@ -143,7 +143,7 @@ void RAMPAdapter::Impl::getScanHeader(size_t index, ScanHeaderStruct& result, bo
         string dissociationMethodName;
         if (!(dissociationMethod.empty()))
             dissociationMethodName = dissociationMethod.name();
-        int len = min(dissociationMethodName.length(), SCANTYPE_LENGTH - 1);
+        int len = min((int)dissociationMethodName.length(), SCANTYPE_LENGTH - 1);
         dissociationMethodName.copy(result.activationMethod, len);
         result.activationMethod[len] = 0; // string.copy does not null terminate
     }
